% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_blue_bright}
\alias{bcg_blue_bright}
\title{Hex code for soft grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_blue_bright
}
\description{
#30C1D7
}
\keyword{datasets}
