% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_blue_true}
\alias{bcg_blue_true}
\title{Hex code for a true blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_blue_true
}
\description{
#295E7E
}
\keyword{datasets}
