% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_colour_manual.R
\name{bcg_colour_manual}
\alias{bcg_colour_manual}
\title{Easily colour your ggplot data using the BCG palette, up to 6 colours}
\usage{
bcg_colour_manual(...)
}
\description{
Easily colour your ggplot data using the BCG palette, up to 6 colours
}
