% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_fill_manual.R
\name{bcg_fill_manual}
\alias{bcg_fill_manual}
\title{Easily fill your ggplot data using the BCG palette, up to 6 colour}
\usage{
bcg_fill_manual(...)
}
\description{
Easily fill your ggplot data using the BCG palette, up to 6 colour
}
