% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_green_1}
\alias{bcg_green_1}
\title{Hex code for the main BCG green}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_green_1
}
\description{
#2ABA75
}
\keyword{datasets}
