% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_green_2}
\alias{bcg_green_2}
\title{Hex code for a darker BCG green}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_green_2
}
\description{
#1A7A55
}
\keyword{datasets}
