% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_green_3}
\alias{bcg_green_3}
\title{Hex code for a darker BCG green}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_green_3
}
\description{
#60DB66
}
\keyword{datasets}
