% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_green_4}
\alias{bcg_green_4}
\title{Hex code for an electric BCG green}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_green_4
}
\description{
#64A64F
}
\keyword{datasets}
