% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_green_bright}
\alias{bcg_green_bright}
\title{Hex code for the main BCG green}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_green_bright
}
\description{
#29BA74
}
\keyword{datasets}
