% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_green_bright_1}
\alias{bcg_green_bright_1}
\title{Hex code for the 1st alternative bright BCG green}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_green_bright_1
}
\description{
#A1E26F
}
\keyword{datasets}
