% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_green_bright_4}
\alias{bcg_green_bright_4}
\title{Hex code for the 4th alternative bright BCG green}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_green_bright_4
}
\description{
#318C46
}
\keyword{datasets}
