% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_green_forest}
\alias{bcg_green_forest}
\title{Hex code for forest green}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_green_forest
}
\description{
#03522D
}
\keyword{datasets}
