% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_green_jade}
\alias{bcg_green_jade}
\title{Hex code for jade}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_green_jade
}
\description{
#197A56
}
\keyword{datasets}
