% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_green_mint}
\alias{bcg_green_mint}
\title{Hex code for mint green}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_green_mint
}
\description{
#3EAD92
}
\keyword{datasets}
