% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_grey_axis}
\alias{bcg_grey_axis}
\title{Hex code for axis grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_grey_axis
}
\description{
#7F7F7F
}
\keyword{datasets}
