% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_grey_background}
\alias{bcg_grey_background}
\title{Hex code for background grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_grey_background
}
\description{
#F2F2F2
}
\keyword{datasets}
