% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_grey_medium}
\alias{bcg_grey_medium}
\title{Hex code for medium grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_grey_medium
}
\description{
#6E6F73
}
\keyword{datasets}
