% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_grey_soft}
\alias{bcg_grey_soft}
\title{Hex code for soft grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_grey_soft
}
\description{
#9A9A9A
}
\keyword{datasets}
