% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_grey_text}
\alias{bcg_grey_text}
\title{Hex code for text grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_grey_text
}
\description{
#575757
}
\keyword{datasets}
