% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_palette.R
\name{bcg_palette}
\alias{bcg_palette}
\title{Output ggplot charts to pdf or png files}
\usage{
bcg_palette(pal = "bright")
}
\arguments{
\item{pal}{Type of palette to pick, "ee" or "multi" or "traffic" or "bright" or "dark" - defaults to bright}
}
\description{
Output ggplot charts to pdf or png files
}
