% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_red_cranberry}
\alias{bcg_red_cranberry}
\title{Hex code for cranberry}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_red_cranberry
}
\description{
#670F31
}
\keyword{datasets}
