% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_red_magenta}
\alias{bcg_red_magenta}
\title{Hex code for magenta}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_red_magenta
}
\description{
#E71C57
}
\keyword{datasets}
