% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_red_orange}
\alias{bcg_red_orange}
\title{Hex code for orange red}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_red_orange
}
\description{
#E15846
}
\keyword{datasets}
