% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_save.R
\name{bcg_save}
\alias{bcg_save}
\title{Output ggplot charts to pdf or png files}
\usage{
bcg_save(..., filename, type)
}
\arguments{
\item{filename}{Name for output file}

\item{type}{Type of chart to output, either "third", "half", "two third", "large", "full", or "all"}
}
\description{
Output ggplot charts to pdf or png files
}
