% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_save_pptx.R
\name{bcg_save_pptx}
\alias{bcg_save_pptx}
\title{Output ggplot charts to powerpoint files}
\usage{
bcg_save_pptx(..., filename, type)
}
\arguments{
\item{filename}{Name for output file}

\item{type}{Type of chart to output, either "halfslide" or "fullslide" or "all"}
}
\description{
Output ggplot charts to powerpoint files
}
