% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_scale_y_continuous.R
\name{bcg_scale_y_continuous}
\alias{bcg_scale_y_continuous}
\title{Reduce the bounds of the y-axis, in BCG style}
\usage{
bcg_scale_y_continuous(...)
}
\description{
Reduce the bounds of the y-axis, in BCG style
}
