% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_theme_classic.R
\name{bcg_theme_classic}
\alias{bcg_theme_classic}
\title{Create a ggplot2 theme in the style of classic, old-school BCG charts}
\usage{
bcg_theme_classic(base_size = 18, base_family = "sans", legend = "bottom")
}
\arguments{
\item{base_size}{Size for text elements. Defaults to 18}

\item{base_family}{Font family for text elements. Defaults to "sans",
indistinguishable from Arial.}

\item{legend}{"bottom" by default.}
}
\description{
Create a ggplot2 theme in the style of classic, old-school BCG charts
}
