% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_theme_modern.R
\name{bcg_theme_modern}
\alias{bcg_theme_modern}
\title{Create a ggplot2 theme in the style of modern, new school BCG charts}
\usage{
bcg_theme_modern(
  base_size = 18,
  base_family = "sans",
  legend = "bottom",
  y_axis = FALSE
)
}
\arguments{
\item{base_size}{Size for text elements. Defaults to 18}

\item{base_family}{Font family for text elements. Defaults to "sans",
indistinguishable from Arial.}

\item{legend}{"bottom" by default.}

\item{y_axis}{FALSE by default. Set TRUE to add a y axis}
}
\description{
Create a ggplot2 theme in the style of modern, new school BCG charts
}
