% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_yellow}
\alias{bcg_yellow}
\title{Hex code for an classic BCG yellow}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_yellow
}
\description{
#DEE341
}
\keyword{datasets}
