% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bcg_yellow_dark}
\alias{bcg_yellow_dark}
\title{Hex code for dark yellow}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcg_yellow_dark
}
\description{
#A8B21C
}
\keyword{datasets}
