% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_titles.R
\name{wrap_titles}
\alias{wrap_titles}
\title{Helper function for use in bcg_save and bcg_save_pptx}
\usage{
wrap_titles(base_plot, subtitle, caption, x, width)
}
\arguments{
\item{base_plot}{The ggplot object to wrap, called internally within save functions}

\item{subtitle}{Subtitle from the ggplot object to wrap, called internally within save functions}

\item{caption}{Caption from the ggplot object to wrap, called internally within save functions}

\item{x}{x-axis label from the ggplot object to wrap, called internally within save functions}

\item{width}{Width in cm}
}
\description{
Helper function for use in bcg_save and bcg_save_pptx
}
