% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_colour_manual.R
\name{bcg_colour_manual}
\alias{bcg_colour_manual}
\title{Easily colour your ggplot data using the BCG palette}
\usage{
bcg_colour_manual(..., pal = "base", reverse = FALSE)
}
\arguments{
\item{pal}{Choose the colour palette, either "base" or "traffic2" or "traffic3". Defaults to "base".}

\item{reverse}{Flip the order of colours in the palette. TRUE or FALSE. Defaults to FALSE.}
}
\description{
Easily colour your ggplot data using the BCG palette
}
