% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_palette.R
\name{bcg_palette}
\alias{bcg_palette}
\title{Output ggplot charts to pdf or png files}
\usage{
bcg_palette(pal = "base")
}
\arguments{
\item{pal}{Type of palette to pick, "base" or "traffic" - defaults to base}
}
\description{
Output ggplot charts to pdf or png files
}
