% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_theme_classic.R
\name{bcg_theme_classic}
\alias{bcg_theme_classic}
\title{Create a ggplot2 theme in the style of classic, old-school BCG charts}
\usage{
bcg_theme_classic(
  base_size = 16,
  background = "white",
  legend = "bottom",
  flipped = FALSE
)
}
\arguments{
\item{base_size}{Size for text elements. Defaults to 16}

\item{background}{"white" by default}

\item{legend}{"bottom" by default.}

\item{flipped}{"FALSE" by default}
}
\description{
Create a ggplot2 theme in the style of classic, old-school BCG charts
}
