% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcg_theme_modern.R
\name{bcg_theme_modern}
\alias{bcg_theme_modern}
\title{Create a ggplot2 theme in the style of modern, new school BCG charts}
\usage{
bcg_theme_modern(
  base_size = 16,
  background = "white",
  legend = "bottom",
  y_axis = FALSE,
  flipped = FALSE
)
}
\arguments{
\item{base_size}{Size for text elements. Defaults to 18}

\item{background}{"white" by default.}

\item{legend}{"bottom" by default.}

\item{y_axis}{FALSE by default. Set TRUE to add a y axis}

\item{flipped}{FALSE by default.}

\item{base_family}{Font family for text elements. Defaults to "sans",
indistinguishable from Arial.}
}
\description{
Create a ggplot2 theme in the style of modern, new school BCG charts
}
